chore: Make the `NcAppNavigationSearch` a fully separated component
